improve stability some e2e
